Parameterize card rendering function
Many parts of the card are based on ratios. To make it easier to try
different card sizes without having to manually recalculate multiple
numbers, I parameterize the rendering function with a single number -
the card height. Based on this number and some hard-coded ratios, I can
derive all the other measurements.