Check for win conditions
Whenever a card is played or a player ends their turn, I check if win
conditions have been met.

Since a lot of code needs to work only in the `Playing` state, I've
introduced monomorphic functor and monad for the `Game` type. These
clean up the `update` function a lot.