Prevent mirror matchups
Previously, the opponent's deck was chosen from _all_ decks. This
allowed mirrors to occur 25% of the time (we have 4 decks for now).

This change will avoid mirrors if there are any other decks available to
play (which should always be the case). If for some reason there was
only a single deck in the game, then a mirror would be allowed.