Convert completely to `Element`
Previously I was using a combination of elm/html and elm-ui. This
converts to use elm-ui everywhere with the following exceptions:

1. The progress bar is using HTML since elm-ui has no progress bar
2. The header is still using HTML since there's no `h1` equivalent in
   elm-ui. I could make it using elements but wanted to hold off on
   that.
3. At the root, I need to use HTML to play nicely with Elm's program.
   Ideally though, I'm only doing a single conversion at the root.

Converting to elm-ui broke the layout a bit. I fixed some padding and
spacing. Since elm-ui doesn't have `hr` equivalents, I changed the
background colors of the main rows to emphasize separation.